Added apa1403
The date they have on their website for this meet is incorrect. The PDF says 11/8/2014 but the link on their site says 11/6/2014. November in 2014 there is no saturday on the 6th but the 8th has a saturday so it is probably the 8th.  Added this meet per request from ronald tarvin. Tee Cummins is also in this meet. All the scripts still need to be done to convert it to kgs and add place and such.