feat: automated lifter details
In the future, we would like to allow lifters to upload information such
as their date of birth and Instagram handle directly on their page. This
will mean they don't have to file an issue with us and we will not need
to triage/create PRs for each one.

The general idea is that a lifter will upload this data, and another
process will handle the creation of one PR every day containing all the
details submitted.

This PR:
* Adds a very basic server-side endpoint for uploading this information