checker: auto-fix errors
There are some errors produced by the checker that could be
automatically fixed if the user desires. For example, if there is an
attempt of `0.0` but we expect an empty column, the checker could be
allowed to fix this itself. The same is true of name conflicts, where a
different accentation could cause errors (and plenty of them in the case
of some Spanish meets).

Instead, the checker should support 3 forms of error:

* "Fatal" errors which must be fixed manually
* Errors which can be automatically resolved if desired
* Warnings which could be ignored if irrelevant

This change:
* Adds a `FixableError` variant and uses it in a few places
* Adds functionality to the checker to actually fix them
* Updates the frontend to handle the new types