refactor: add editor with less validation
As discussed in review, the use of `Serialize` for `Entry` means that
CSVs will always be returned in the same header order, no matter of
their original header order. What we'd prefer is to have a more basic
`Editor` type of some form that does little validation and just edits
in-place.

This adds that struct with some tests and validation, and uses it in the
checker page. This doesn't change the behaviour at all.