fix: only show check results in `Mode::Check`
Showing the check results when running in `Mode::Fix` is a little
confusing, because we may have already fixed some of them. Instead,
let's clear the results and force the user to run `Check` again to get
the updated results.

In the future, the `checker` should fix everything it can and then run
another `Check` pass against the result to return the new (unfixed)
errors.